+++ ng2-components/ng2-activiti-tasklist/src/assets/apps-list.mock.ts
